requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.11.0.post1
fonttools==4.58.5
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
jiter==0.10.0
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
openai==1.96.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
zipp==3.23.0
